some refactoring :)
