+++ go/libraries/doltcore/sqle/sqlselect_test.go
@@ -24,12 +24,12 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/dolthub/dolt/go/libraries/doltcore/doltdb"
+	"github.com/dolthub/dolt/go/libraries/doltcore/doltdocs"
 	"github.com/dolthub/dolt/go/libraries/doltcore/dtestutils"
 	"github.com/dolthub/dolt/go/libraries/doltcore/envtestutils"
 	"github.com/dolthub/dolt/go/libraries/doltcore/row"
 	"github.com/dolthub/dolt/go/libraries/doltcore/schema"
 	. "github.com/dolthub/dolt/go/libraries/doltcore/sql/sqltestutil"
-	"github.com/dolthub/dolt/go/libraries/doltcore/sqle/docsTable"
 	"github.com/dolthub/dolt/go/libraries/doltcore/sqle/dtables"
 	"github.com/dolthub/dolt/go/store/types"
 )
@@ -1465,15 +1465,15 @@ var systemTableSelectTests = []SelectTest{
 	{
 		Name: "select from dolt_docs",
 		AdditionalSetup: CreateTableFn("dolt_docs",
-			docsTable.DoltDocsSchema,
-			NewRowWithSchema(docsTable.DoltDocsSchema,
+			doltdocs.DoltDocsSchema,
+			NewRowWithSchema(doltdocs.DoltDocsSchema,
 				types.String("LICENSE.md"),
 				types.String("A license")),
 		),
 		Query: "select * from dolt_docs",
-		ExpectedRows: ToSqlRows(CompressSchema(docsTable.DoltDocsSchema),
+		ExpectedRows: ToSqlRows(CompressSchema(doltdocs.DoltDocsSchema),
 			NewRow(types.String("LICENSE.md"), types.String("A license"))),
-		ExpectedSchema: CompressSchema(docsTable.DoltDocsSchema),
+		ExpectedSchema: CompressSchema(doltdocs.DoltDocsSchema),
 	},
 	{
 		Name: "select from dolt_query_catalog",
